Forms dir was translated on tsx
