+++ frontend/src/components/Forms/PasswordVisibilityButton.tsx
