+++ frontend/src/components/Forms/UpdateAccountForm.tsx
@@ -9,21 +9,23 @@ import { object } from 'yup';
 import Button from 'react-bootstrap/Button';
 import Form from 'react-bootstrap/Form';
 
-import { actions as userActions } from '../../slices/userSlice.js';
+import { actions as userActions } from '../../slices/userSlice';
 import routes from '../../routes';
 import { email, username } from '../../utils/validationSchemas';
-import FormAlert from './FormAlert.jsx';
+import FormAlert from './FormAlert';
+import { RootReducerType } from 'src/types/slices';
+import { TypeInitialFormState } from 'src/types/components';
 
 function UpdateAccountForm() {
   const { t: tPS } = useTranslation('translation', {
     keyPrefix: 'profileSettings',
   });
   const { t } = useTranslation();
-  const emailRef = useRef();
-  const usernameRef = useRef();
+  const emailRef = useRef<HTMLInputElement>(null);
+  const usernameRef = useRef<HTMLInputElement>(null);
   const dispatch = useDispatch();
-  const userInfo = useSelector((state) => state.user.userInfo);
-  const initialFormState = { state: 'initial', message: '' };
+  const userInfo = useSelector((state: RootReducerType) => state.user.userInfo);
+  const initialFormState: TypeInitialFormState = { state: 'initial', message: '' };
   const [formState, setFormState] = useState(initialFormState);
 
   const validationSchema = object().shape({
@@ -68,8 +70,8 @@ function UpdateAccountForm() {
           err.response?.status === 400 &&
           Array.isArray(err.response?.data?.errs?.message)
         ) {
-          const errResponseMessages = err.response.data.errs.message;
-          errResponseMessages.forEach((e) => {
+          const errResponseMessages: string[] = err.response.data.errs.message;
+          errResponseMessages.forEach((e): never | void => {
             switch (e) {
               case 'usernameIsUsed':
                 actions.setFieldError(
@@ -119,7 +121,7 @@ function UpdateAccountForm() {
         <Form.Control
           ref={usernameRef}
           autoComplete="username"
-          isInvalid={formik.touched.username && formik.errors.username}
+          isInvalid={!!(formik.touched.username && formik.errors.username)}
           name="username"
           onBlur={formik.handleBlur}
           onChange={formik.handleChange}
